# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Lead Scout API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  owner: 'owner',
  manager: 'manager',
} as const;

export interface User {
  id: number;
  name: string;
  login: string;
  role: UserRole;
  active: boolean;
}

export interface LoginInput {
  login: string;
  password: string;
}

export interface Session {
  authenticated: boolean;
  user?: User | null;
}

export interface Country {
  code: string;
  name: string;
  enabled: boolean;
}

export interface CountryInput {
  code: string;
  name: string;
}

export interface Contact {
  type: string;
  value: string;
  url: string;
}

/**
 * Whether the source provided a company website URL.
 */
export type LeadWebsiteStatus = typeof LeadWebsiteStatus[keyof typeof LeadWebsiteStatus];


export const LeadWebsiteStatus = {
  present: 'present',
  missing: 'missing',
} as const;

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  new: 'new',
  claimed: 'claimed',
  contacted: 'contacted',
  replied: 'replied',
  rejected: 'rejected',
  no_reply: 'no_reply',
  deal: 'deal',
} as const;

export type WebsiteAuditStatus = typeof WebsiteAuditStatus[keyof typeof WebsiteAuditStatus];


export const WebsiteAuditStatus = {
  checked: 'checked',
  unavailable: 'unavailable',
  not_provided: 'not_provided',
} as const;

export type WebsiteAuditCheckStatus = typeof WebsiteAuditCheckStatus[keyof typeof WebsiteAuditCheckStatus];


export const WebsiteAuditCheckStatus = {
  pass: 'pass',
  warn: 'warn',
  fail: 'fail',
  unknown: 'unknown',
} as const;

export interface WebsiteAuditCheck {
  key: string;
  label: string;
  status: WebsiteAuditCheckStatus;
  evidence: string;
}

export interface WebsiteAudit {
  status: WebsiteAuditStatus;
  /** @nullable */
  url: string | null;
  /** @nullable */
  finalUrl: string | null;
  /** @nullable */
  checkedAt: string | null;
  /** @nullable */
  responseTimeMs: number | null;
  /** @nullable */
  statusCode: number | null;
  /** @nullable */
  pageSizeKb: number | null;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  qualityScore: number | null;
  /** @nullable */
  title: string | null;
  /** @nullable */
  metaDescription: string | null;
  /** @nullable */
  hasMobileViewport: boolean | null;
  checks: WebsiteAuditCheck[];
  /** @nullable */
  error: string | null;
}

export type ScoreFactorCategory = typeof ScoreFactorCategory[keyof typeof ScoreFactorCategory];


export const ScoreFactorCategory = {
  market: 'market',
  contactability: 'contactability',
  opportunity: 'opportunity',
  website: 'website',
} as const;

export interface ScoreFactor {
  key: string;
  label: string;
  category: ScoreFactorCategory;
  points: number;
  maxPoints: number;
  evidence: string;
}

export interface Lead {
  id: number;
  name: string;
  country: string;
  city: string;
  /** @nullable */
  address: string | null;
  industry: string;
  /** @nullable */
  website?: string | null;
  /** Whether the source provided a company website URL. */
  websiteStatus: LeadWebsiteStatus;
  status: LeadStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  scoreReasons: string[];
  issues: string[];
  reviewsCount: number;
  /** @nullable */
  rating?: number | null;
  branchesCount: number;
  contacts: Contact[];
  source: string;
  /**
     * Direct company profile URL at the source
     * @nullable
     */
  sourceUrl: string | null;
  websiteAudit?: WebsiteAudit | null;
  scoreBreakdown: ScoreFactor[];
  scoreVersion: string;
  assignee?: User | null;
  /** @nullable */
  note?: string | null;
  updatedAt: string;
}

export interface LeadSearchInput {
  country: string;
  /** Empty means search across the supported 2GIS cities in the selected country. */
  city: string;
  /** Empty means use a broad organization search rather than a selected industry. */
  industry: string;
  /** Include companies already shown to this user in earlier searches. */
  showPreviouslyFound?: boolean;
}

export type LeadUpdateStatus = typeof LeadUpdateStatus[keyof typeof LeadUpdateStatus];


export const LeadUpdateStatus = {
  new: 'new',
  claimed: 'claimed',
  contacted: 'contacted',
  replied: 'replied',
  rejected: 'rejected',
  no_reply: 'no_reply',
  deal: 'deal',
} as const;

export interface LeadUpdate {
  status?: LeadUpdateStatus;
  note?: string;
}

export interface StageCount {
  status: string;
  count: number;
}

export interface Activity {
  id: number;
  text: string;
  at: string;
}

export interface Dashboard {
  available: number;
  inWork: number;
  replies: number;
  deals: number;
  stages: StageCount[];
  recent: Activity[];
}

export type ListLeadsParams = {
status?: string;
assignedToMe?: boolean;
};

export type ListLeadArchiveParams = {
/**
 * Search all clients ever found by name, city, industry, or website.
 */
query?: string;
/**
 * Filter archived clients by pipeline status.
 */
status?: string;
};