# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Lead Scout API
 * OpenAPI spec version: 0.1.0
 */
import type { Contact } from './contact';
import type { LeadStatus } from './leadStatus';
import type { LeadWebsiteStatus } from './leadWebsiteStatus';
import type { ScoreFactor } from './scoreFactor';
import type { User } from './user';
import type { WebsiteAudit } from './websiteAudit';

export interface Lead {
  id: number;
  name: string;
  country: string;
  city: string;
  /** @nullable */
  address: string | null;
  industry: string;
  /** @nullable */
  website?: string | null;
  /** Whether the source provided a company website URL. */
  websiteStatus: LeadWebsiteStatus;
  status: LeadStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  scoreReasons: string[];
  issues: string[];
  reviewsCount: number;
  /** @nullable */
  rating?: number | null;
  branchesCount: number;
  contacts: Contact[];
  source: string;
  /**
     * Direct company profile URL at the source
     * @nullable
     */
  sourceUrl: string | null;
  websiteAudit?: WebsiteAudit | null;
  scoreBreakdown: ScoreFactor[];
  scoreVersion: string;
  assignee?: User | null;
  /** @nullable */
  note?: string | null;
  updatedAt: Date;
}