# lib/api-zod/src/generated/types/leadSearchInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Lead Scout API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadSearchInput {
  country: string;
  /** Empty means search across the supported 2GIS cities in the selected country. */
  city: string;
  /** Empty means use a broad organization search rather than a selected industry. */
  industry: string;
  /** Include companies already shown to this user in earlier searches. */
  showPreviouslyFound?: boolean;
}